refactoring file naming conventions to kabob-case
