+++ src/components/fields/base/option-fields/delete.vue
