+++ src/components/fields/base/text-fields/delete.vue
