--- src/layouts/simpleLayout.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-  <slot/>
-  </div>
-</template>
