+++ src/main.js
@@ -12,17 +12,17 @@ import axios from '@/axios'
 import '@/assets/css/styles.css'
 
 /* Layouts */
-import simpleLayout from '@/layouts/simpleLayout'
-import appLayout from '@/layouts/appLayout'
+import simpleLayout from '@/layouts/simple-layout'
+import appLayout from '@/layouts/app-layout'
 
 /* Components */
 import baseModal from '@/components/global/modals/base'
 import statCard from '@/components/global/cards/stat-card'
-import checkboxSingle from '@/components/global/forms/base_fields/checkbox-single'
-import selectField from '@/components/global/forms/base_fields/select-field'
-import textField from '@/components/global/forms/base_fields/text-field'
-import vTextField from '@/components/global/forms/validation_fields/text-field'
-import vSelectField from '@/components/global/forms/validation_fields/select-field'
+import checkboxSingle from '@/components/global/forms/base-fields/checkbox-single'
+import selectField from '@/components/global/forms/base-fields/select-field'
+import textField from '@/components/global/forms/base-fields/text-field'
+import vTextField from '@/components/global/forms/validation-fields/text-field'
+import vSelectField from '@/components/global/forms/validation-fields/select-field'
 
 /* -----------------------------------------------
   Authentication
